auth-service/requirements.txt
aio-pika==9.4.3
aiormq==6.8.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
loguru==0.7.2
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
pamqp==3.3.0
pika==1.3.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-env==1.1.5
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==13.0
yarl==1.9.4
